+++ src/app/models/member.ts
@@ -1,4 +1,5 @@
 export class Member {
+  avatarHash: string;
   constructor(public id?: string,
               public fullName?: string,
               public username?: string) {